pyproject.toml
[project]
name = "anki-generator"
version = "0.1.0"
description = "Generate Anki cards from Chinese words using DeepSeek API"
requires-python = ">=3.11"
dependencies = [
    "openai>=1.0.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
anki-gen = "src.main:main"
